refactoring, bug fixes

+++ src/main/java/com/lb_calc_web/model/LC.java
@@ -15,7 +15,7 @@ import java.util.Objects;
 public class LC {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private int height;
@@ -57,11 +57,11 @@ public class LC {
         this.bottomFrame = bottomFrame;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
